+++ src/geometry/ray.rs
@@ -1,4 +1,5 @@
 use crate::geometry::feature::IntoTypeValue;
+use crate::geometry::RawFeatureType;
 use crate::math::RawVector;
 use rapier::geometry::{ColliderHandle, RayIntersection};
 use wasm_bindgen::prelude::*;
@@ -16,8 +17,8 @@ impl RawRayIntersection {
         self.0.toi
     }
 
-    pub fn featureType(&self) -> Option<u32> {
-        self.0.feature.into_type().into()
+    pub fn featureType(&self) -> RawFeatureType {
+        self.0.feature.into_type()
     }
 
     pub fn featureId(&self) -> Option<u32> {
@@ -45,8 +46,8 @@ impl RawRayColliderIntersection {
         self.inter.toi
     }
 
-    pub fn featureType(&self) -> Option<u32> {
-        self.inter.feature.into_type().into()
+    pub fn featureType(&self) -> RawFeatureType {
+        self.inter.feature.into_type()
     }
 
     pub fn featureId(&self) -> Option<u32> {